refactor: items renamed to products
