+++ server/src/models/index.model.ts
@@ -1,7 +1,7 @@
 export { default as Category } from './category.model';
 export { default as Color } from './color.model';
 export { default as Customers } from './customer.model';
-export { default as Item } from './item.model';
+export { default as Item } from './product.model';
 export { default as Order } from './order.model';
 export { default as Session } from './session.model';
 export { default as Stock } from './stock.model';
